Note It
A simple note taking app for iPhone built with Swift 3.

I'll be updating the app in the coming days and weeks. This is more or
less what you can expect for a roadmap.

-[x] Add support for rich-text-editing.
-[] Add support for groups/notebooks.
-[] Add trash/deleted group which self-purges.
-[] Add support for sharing.
-[] Add support for cloud storage (iCloud and/or Dropbox.)

This first version of the application is based in large part on a
tutorial by Todd Perkins (@askTodd) on Lynda.com
[https://www.lynda.com/Swift-tutorials/Next-steps/504400/579555-4.html].
 It's a great tutorial by a great instructor from a great service. I
totally reccomend it.

This app will not be released on the App Store so if you'd like to try
it out, feel free to download the latest version and load it onto your
device using Xcode.

Thanks for checking out my first project here on GitHub.